Updating highsocre

+++ Assets/PlayFabSDK/Client/PlayFabClientModels.cs
@@ -6165,7 +6165,7 @@ namespace PlayFab.ClientModels
         /// <summary>
         /// statistic value for the player
         /// </summary>
-        public int Value;
+        public float Value;
         /// <summary>
         /// for updates to an existing statistic value for a player, the version of the statistic when it was loaded. Null when
         /// setting the statistic value for the first time.
@@ -6183,7 +6183,7 @@ namespace PlayFab.ClientModels
         /// <summary>
         /// statistic value for the player
         /// </summary>
-        public int Value;
+        public float Value;
         /// <summary>
         /// for updates to an existing statistic value for a player, the version of the statistic when it was loaded
         /// </summary>
